003_azure_security/user_authentication_authorization/requirements.txt
python-dotenv
jupyterlab
msal